+++ packages/core/src/components/annotation/annotation-gebiedsaanwijzing/annotation-gebiedsaanwijzing.tsx
@@ -1,8 +1,10 @@
-import { Component, ComponentInterface, Prop, Event, h, EventEmitter, Fragment } from "@stencil/core";
-import { AnnotationActiveChangeEvent, AnnotationDiff, AnnotationWijzigactie } from "../annotation.interfaces";
+import { Component, ComponentInterface, Prop, Event, h, EventEmitter, Fragment, Element } from "@stencil/core";
+import { AnnotationActiveChangeEvent, AnnotationWijzigactie } from "../annotation.interfaces";
 import { AnnotationBody } from "../annotation-body";
-import { AnnotationDiffRenderer } from "../annotation-diff-renderer";
 import { AnnotationGewijzigdeLocatie } from "../annotation-gewijzigde-locatie";
+import { watcher } from "../annotation-watcher";
+import { AnnotationSymbolSlot } from "../annotation-symbol-slot";
+import { RenvooiValue } from "../../renvooi/renvooi.interfaces";
 
 /**
  * @slot symbool - Een optionele afbeelding die de annotatie symboliseert.
@@ -41,17 +43,33 @@ export class AnnotationGebiedsaanwijzing implements ComponentInterface {
    * De naam van de gebiedsaanwijzing.
    */
   @Prop()
-  naam?: AnnotationDiff | string;
+  naam?: RenvooiValue | string;
+
+  @Element()
+  private host!: HTMLDsoAnnotationGebiedsaanwijzingElement;
+
+  private watcher = watcher(this.host);
+
+  connectedCallback(): void {
+    this.watcher.watch();
+  }
+
+  disconnectedCallback(): void {
+    this.watcher.unwatch();
+  }
 
   render() {
+    const hasSymbool = this.watcher.hasSymbool();
+
     return (
       <AnnotationBody
+        symbol={hasSymbool ? <AnnotationSymbolSlot /> : undefined}
         active={this.active}
         dsoActiveChange={this.dsoActiveChange}
         title={
           <>
             <span class="content">
-              <AnnotationDiffRenderer value={this.naam} />
+              <dso-renvooi value={this.naam} />
             </span>
             {this.gewijzigdeLocatie && <AnnotationGewijzigdeLocatie />}
           </>