+++ packages/core/src/components/annotation/annotation-symbol-slot.tsx
@@ -1,0 +1,3 @@
+import { h, FunctionalComponent } from "@stencil/core";
+
+export const AnnotationSymbolSlot: FunctionalComponent = () => <slot name="symbool" />;