+++ packages/core/src/components/renvooi/renvooi.interfaces.ts
@@ -1,0 +1,1 @@
+export type RenvooiValue = { toegevoegd: string } | { was: string; wordt: string } | { verwijderd: string } | string;