+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Added
+* Annotation Component: Doorontwikkeling ([#2691](https://github.com/dso-toolkit/dso-toolkit/issues/2691))
+
 ### Fixed
 * Form Buttons: Button stretches within container ([#2721](https://github.com/dso-toolkit/dso-toolkit/issues/2721))
 * Rich Content: Nested list remove block end margin ([#2709](https://github.com/dso-toolkit/dso-toolkit/issues/2709))